bot/requirements.txt
python-telegram-bot>=21.0
openai>=1.0
python-dotenv>=1.0
requests>=2.31
markdown>=3.5
